% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{app_server}
\alias{app_server}
\title{serveur Picross}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{inputs}

\item{output}{output}

\item{session}{session}
}
\value{
serveur
}
\description{
serveur Picross
}
\examples{
exemple
}
