% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_cancer.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{Data breast cancer}
\format{
An object of class \code{data.frame} with 569 rows and 33 columns.
}
\usage{
cancer
}
\description{
A dataset containing the diagnosis and 32 others variables
}
\keyword{datasets}
