% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrer_reduire.R
\name{centr_reduc_dataframe}
\alias{centr_reduc_dataframe}
\title{Function centr_reduc_dataframe}
\usage{
centr_reduc_dataframe(X)
}
\arguments{
\item{X}{a dataframe cantaining the variables in the model}
}
\value{
Returns the new dataframe with standarized variables
}
\description{
centr_reduc_dataframe appplies centr_reduc_vecteur on the dataframe
}
