% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrer_reduire.R
\name{centr_reduc_vecteur}
\alias{centr_reduc_vecteur}
\title{Function centr_reduc_vecteur}
\usage{
centr_reduc_vecteur(x)
}
\arguments{
\item{x}{it's a param}
}
\value{
Returns standarized variables
}
\description{
centr_reduc_vecteur function allows us to standarize our variables.
}
