% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{classe_appartenance}
\alias{classe_appartenance}
\title{Function class_belonging}
\usage{
classe_appartenance(x)
}
\arguments{
\item{x}{it's a parameter}
}
\value{
Returns 0 or 1
}
\description{
class_belonging function allows to calculate the class of belonging : 1 if x>0 otherwise 0
}
