% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_SGD.R
\name{gradient_stochastique}
\alias{gradient_stochastique}
\title{Function gradient_stochastique}
\usage{
gradient_stochastique(
  X,
  y,
  theta,
  epsilon,
  learning_rate,
  max_iter,
  batch_size
)
}
\arguments{
\item{X}{a numetric matrix}

\item{y}{a numeric vector}

\item{theta}{is a gradient coefficient}

\item{epsilon}{The parametre that tells us when we converge}

\item{learning_rate}{a double value}

\item{max_iter}{a integer}

\item{batch_size}{Select a sample of the dataframe(batch_size=1:online,batch_size=nrow(data):batch,1<batch_size<nrow(data):mini-batch )}
}
\value{
a vector theta which is descent gradient coefficients
}
\description{
Gradient_stochastique function to calcul the simple gradient.
}
