% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{nonogramecreate}
\alias{nonogramecreate}
\title{Create a Nonogram Puzzle}
\usage{
nonogramecreate(n, diff)
}
\arguments{
\item{n}{The size of the Nonogram puzzle grid (number of rows/columns).}

\item{diff}{The difficulty level of the puzzle. Options are "Easy" and "Hard".}
}
\value{
A list containing the following components:
  - \code{matindichori} A matrix representing horizontal indications for each row.
  - \code{matindicvert} A matrix representing vertical indications for each column.
  - \code{X} The generated Nonogram grid.
  - \code{compteurH} A vector containing counts of filled cells for each row.
  - \code{compteurV} A vector containing counts of filled cells for each column.
}
\description{
This function generates a Nonogram puzzle with specified dimensions and difficulty level.
}
\details{
The Nonogram puzzle is generated based on a random matrix of 0s and 1s, where
  0 represents an empty cell and 1 represents a filled cell. The function also generates
  horizontal and vertical indications based on the filled cells. The difficulty level 
  determines whether additional constraints are applied to generate the puzzle.
}
\examples{
# Generate an easy Nonogram puzzle with a 5x5 grid
nonogramecreate(5, "Easy")

# Generate a hard Nonogram puzzle with a 10x10 grid
nonogramecreate(10, "Hard")

}
