% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_Server.R
\name{Piccross_R}
\alias{Piccross_R}
\title{Run the Picross Shiny App}
\usage{
Piccross_R()
}
\description{
This function launches the Picross Shiny App.
}
\details{
The Picross Shiny App allows users to play Picross puzzles interactively.
}
