% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict}
\alias{predict}
\title{Function predict}
\usage{
predict(objet, newdata, type)
}
\arguments{
\item{objet}{Object is a Fit S3 object}

\item{newdata}{a dataframe the same type we use in Fit}

\item{type}{Type is qualitative variable with two modalities : class and probability}
}
\value{
Returns a list of prediction or probability
}
\description{
Predict function allows us to do our prediction
}
