% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_SGD.R
\name{prediction}
\alias{prediction}
\title{Function prediction}
\usage{
prediction(z)
}
\arguments{
\item{z}{it's matrix calculation between X and theta}
}
\value{
Returns a vector
}
\description{
prediction function allows to calculate the sigmoid of a value
}
