% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Reg_Logistique}
\alias{print.Reg_Logistique}
\title{Function print.Reg.Logistique}
\usage{
\method{print}{Reg_Logistique}(x, ...)
}
\arguments{
\item{x}{S3 type Reg_Logistique}

\item{...}{unknown params}
}
\description{
print.Reg.Logistique allows us to print logistic function
}
