% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rotate90}
\alias{rotate90}
\title{Rotate a Matrix by 90 Degrees}
\usage{
rotate90(mat)
}
\arguments{
\item{mat}{A matrix to be rotated.}
}
\value{
The input matrix rotated by 90 degrees.
}
\description{
This function rotates a given matrix by 90 degrees clockwise.
}
\examples{
mat <- matrix(1:4, nrow = 2)
rotate90(mat)
# Output:
#      [,1] [,2]
# [1,]    3    1
# [2,]    4    2

}
