% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrer_reduire.R
\name{save.sd}
\alias{save.sd}
\title{Fonction save.sd}
\usage{
save.sd(x)
}
\arguments{
\item{x}{corresponds to the column of the dataframe}
}
\description{
Returns the standard deviation for each column
}
