% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_SGD.R
\name{SGD_paralle}
\alias{SGD_paralle}
\title{Function SGD_paralle}
\usage{
SGD_paralle(data, theta, k, epsilon, learning_rate, max_iter, batch_size)
}
\arguments{
\item{data}{is a dataframe that containes X and y}

\item{theta}{is a gradient coefficient}

\item{k}{is the nubmer of cores to want use}

\item{epsilon}{The parametre that tells us when we converge}

\item{learning_rate}{a double value}

\item{max_iter}{an integer}

\item{batch_size}{Select a sample of the dataframe(batch_size=1:online,batch_size=nrow(data):batch,1<batch_size<nrow(data):mini-batch )}
}
\value{
a combination of k descent gradient vectors
}
\description{
SGD_paralle function is the same as gradient_stochastique function but with parallelisation
}
