% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.Reg_Logistique}
\alias{summary.Reg_Logistique}
\title{Function summary.Reg.Logistique}
\usage{
\method{summary}{Reg_Logistique}(object, ...)
}
\arguments{
\item{object}{S3 type Reg_Logistique}

\item{...}{other params}
}
\description{
summary.Reg.Logistique allows us to print logistic function
}
