% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valeur_manquante.R
\name{valeurs_manquantes}
\alias{valeurs_manquantes}
\title{Function valeurs_manquantes}
\usage{
valeurs_manquantes(x)
}
\arguments{
\item{x}{corresponds to each row}
}
\value{
Returns modified rows
}
\description{
valeurs_manquantes function allows you to replace or delete a certain column which contains several NA values.
}
