% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valeur_manquante.R
\name{valeurs_manquantes_dataframe}
\alias{valeurs_manquantes_dataframe}
\title{Function valeurs_manquantes_dataframe}
\usage{
valeurs_manquantes_dataframe(X)
}
\arguments{
\item{X}{it's a dataframe}
}
\value{
Returns the new dataframe without missing values
}
\description{
valeurs_manquantes_dataframe function allows to replace or delete a certain column that containes multiple NA values or blank values in the dataframe and returns a new one.
}
