% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrer_reduire.R
\name{centr_reduc_dataframe}
\alias{centr_reduc_dataframe}
\title{Function centr_reduc_dataframe}
\usage{
centr_reduc_dataframe(X)
}
\arguments{
\item{X}{it's a dataframe}
}
\value{
Returns the new dataframe with standarized variables
}
\description{
centr_reduc_dataframe function allows us to standarize our variables in the dataframe.
}
