% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_SGD.R
\name{cost_derive}
\alias{cost_derive}
\title{Function cost_derive}
\usage{
cost_derive(theta, X, y)
}
\arguments{
\item{theta}{is gradient parameter}

\item{X}{a  numeric matrix of features}

\item{y}{a numeric matrix to target variable}
}
\value{
Returns gradient
}
\description{
cost_derive function is the derivation of the log-loss function
}
