% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{fit}
\alias{fit}
\title{Function Fit}
\usage{
fit(
  formula,
  data,
  epsilon = 1e-04,
  learning_rate = 0.01,
  max_iter = 1000,
  batch_size = 1,
  ncores
)
}
\arguments{
\item{formula}{Formula helps us to select the target variable and features}

\item{data}{select a Dataframe}

\item{epsilon}{parameter that tells us when we converge}

\item{learning_rate}{a double value}

\item{max_iter}{an integer}

\item{batch_size}{Select a sample of the dataframe(batch_size=1:online ; batch_size=nrow(data):batch ; 1<batch_size<nrow(data):mini-batch )}

\item{ncores}{Number of processors}
}
\value{
S3 object type Reg_Logistique
}
\description{
Fit function allows us to calculate the coefficients with a Stochastic gradient descent.
}
