% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_SGD.R
\name{gradient_stochastique}
\alias{gradient_stochastique}
\title{Function gradient_stochastique}
\usage{
gradient_stochastique(
  X,
  y,
  theta,
  epsilon,
  learning_rate,
  max_iter,
  batch_size
)
}
\arguments{
\item{X}{a numeric matrix of features}

\item{y}{a numeric matrix to target variable}

\item{theta}{a numeric vector.It includes the same number of data as the column number of X. It is initiated randomly. Theta allows us to calculate the gradients.}

\item{epsilon}{a numerical value that allows us to stop when calculating gradients.}

\item{learning_rate}{numeric value which determines the speed of convergence of the process}

\item{max_iter}{a numeric value to allows stop the gradient descent if we don't converge.}

\item{batch_size}{a number to allow selection of a sample from the database}
}
\value{
a vector theta which is descent gradient coefficients
}
\description{
It's a function to calculate the stochasitic gradient descent.
Batch size corresponds to the sampling mode. If it's equal to 1 then we are doing online. Included between 1 and number of lines, it's a mini batch.
And if it's equal to the number of lines then it's a batch.
}
