% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_shiny.R
\name{lauch_picross}
\alias{lauch_picross}
\title{lancement du jeu de Picross}
\usage{
lauch_picross(ui, server)
}
\arguments{
\item{ui}{ui}

\item{server}{server}
}
\value{
lance la grille de picross avec solution
}
\description{
lancement du jeu de Picross
}
\examples{
exemple
}
