% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_SGD.R
\name{LL}
\alias{LL}
\title{Function LL}
\usage{
LL(theta, X, y)
}
\arguments{
\item{theta}{is gradient parametre}

\item{X}{a  numeric matrix}

\item{y}{a numeric vector}
}
\value{
Returns cost value
}
\description{
LL it's a  log-loss function
}
