% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict}
\alias{predict}
\title{Function predict}
\usage{
predict(objet, newdata, type)
}
\arguments{
\item{objet}{Object is a Fit S3 object}

\item{newdata}{a same dataframe that use in Fit, but without the target variable}

\item{type}{Type is qualitative variable with two modalities : class and posterior}
}
\value{
Returns a list of prediction (0/1) or probability
}
\description{
Predict is a function that allows us to make our prediction
}
