% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_SGD.R
\name{SGD_paralle}
\alias{SGD_paralle}
\title{Function SGD_paralle}
\usage{
SGD_paralle(data, theta, k, epsilon, learning_rate, max_iter, batch_size)
}
\arguments{
\item{data}{a numeric matrix of features}

\item{theta}{a numeric vector.It includes the same number of data as the column number of X. It is initiated randomly. Theta allows us to calculate the gradients.}

\item{k}{allows to split our data set according to the hearts which were entered.}

\item{epsilon}{a numerical value that allows us to stop when calculating gradients.}

\item{learning_rate}{numeric value which determines the speed of convergence of the process}

\item{max_iter}{a numeric value to allows stop the gradient descent if we don't converge.}

\item{batch_size}{a number to allow selection of a sample from the database}
}
\value{
a combination of k descent gradient vectors
}
\description{
SGD_paralle function is the same as gradient_stochastique function but with parallelisation
Batch size corresponds to the sampling mode. If it's equal to 1 then we are doing online. Included between 1 and number of lines, it's a mini batch.
And if it's equal to the number of lines then it's a batch.
}
