% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valeur_manquante.R
\name{valeurs_manquantes}
\alias{valeurs_manquantes}
\title{Function valeurs_manquantes}
\usage{
valeurs_manquantes(x)
}
\arguments{
\item{x}{it a param}
}
\value{
Returns modified rows
}
\description{
valeurs_manquantes function allows to replace or delete a certain column that containes multiple NA values or blank values
}
