% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valeur_manquante.R
\name{valeurs_manquantes_dataframe}
\alias{valeurs_manquantes_dataframe}
\title{Function valeurs_manquantes_dataframe}
\usage{
valeurs_manquantes_dataframe(X)
}
\arguments{
\item{X}{a dataframe cantaining the variables in the model}
}
\value{
Returns the new dataframe without missing values
}
\description{
valeurs_manquantes_dataframe applies valeurs_manquantes to the dataframe
}
